more adaptor -> adapter renames
